refactor to hamilton.plugins
