--- examples/experiment_management/pyproject.toml
@@ -1,22 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "hamilton-experiments"
-version = "0.0.1"
-dependencies = [
-    "python-multipart",
-    "fastui",
-    "fastapi",
-    "sf-hamilton[visualization]",
-    "pydantic",
-    "pandas",
-    "uvicorn",
-]
-
-[project.scripts]
-hexperiments = "hamilton_experiments.__main__:main"
-
-[tool.hatch.build.targets.wheel]
-packages = ["hamilton_experiments/"]
