+++ hamilton/plugins/h_experiments/__init__.py
@@ -1,0 +1,5 @@
+from hamilton.plugins.h_experiments.hook import ExperimentTracker
+
+__all__ = [
+    "ExperimentTracker",
+]
